requirements.txt
ConfigArgParse==1.5.3
flax==0.6.8
matplotlib==3.7.1
numpy==1.23.5
optax==0.1.4
PyYAML==6.0
scikit-image==0.20.0
scipy==1.10.1
tensorflow-cpu==2.12.0
tqdm==4.65.0
trimesh==3.21.5